pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "text2sql-grpo-azure-ml"
version = "0.1.0"
description = "Enterprise-grade GRPO pipeline for Text-to-SQL schema generalisation on Azure ML"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
keywords = ["text-to-sql", "grpo", "llm", "azure-ml", "reinforcement-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.2.0",
    "transformers>=4.40.0",
    "trl>=0.8.6",
    "datasets>=2.18.0",
    "sqlglot>=23.0.0",
    "sqlparse>=0.5.0",
    "mlflow>=2.12.0",
    "azure-ai-ml>=1.15.0",
    "azure-identity>=1.16.0",
    "pandas>=2.2.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "loguru>=0.7.0",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
demo = ["streamlit>=1.33.0"]
dev = [
    "pytest>=8.1.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "black>=24.3.0",
    "isort>=5.13.0",
]

[project.urls]
Homepage = "https://github.com/sriksmachi/text2sql-grpo-azure-ml"
Repository = "https://github.com/sriksmachi/text2sql-grpo-azure-ml"
Issues = "https://github.com/sriksmachi/text2sql-grpo-azure-ml/issues"

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "BLE", "RUF"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]
